fixup! [TEST] Use std::to_chars() instead of std::to_string()
ubuntu 20.04 gcc 9

error log
```
../src/jdlib/jdiconv.cpp: In member function ‘const char* JDLIB::Iconv::convert(char*, int, int&)’:
../src/jdlib/jdiconv.cpp:198:93: error: no matching function for call to ‘to_chars(char [16], char*, const char32_t&, int)’
  198 |                         const auto result = std::to_chars( uni_str, uni_str + 15, unich, 10 );
      |                                                                                             ^
```

+++ src/jdlib/jdiconv.cpp
@@ -8,9 +8,10 @@
 #include "misccharcode.h"
 #include "miscmsg.h"
 
+#include <charconv>
+#include <cstdint>
 #include <cstdio>
 #include <cstring>
-#include <charconv>
 #include <errno.h>
 
 
@@ -195,7 +196,8 @@ const char* Iconv::convert( char* str_in, int size_in, int& size_out )
                         }
 
                         char uni_str[16];
-                        const auto result = std::to_chars( uni_str, uni_str + 15, unich, 10 );
+                        const auto result = std::to_chars( uni_str, uni_str + 15,
+                                                           static_cast<std::uint32_t>( unich ), 10 );
                         const auto uni_str_size = result.ptr - uni_str;
 #ifdef _DEBUG
                         std::cout << "utf32 = " << unich << " byte = " << byte << std::endl;
